feat: Refactor AppSidebar to include logout functionality and remove standalone LogoutButton component
