+++ src/components/app-sidebar.tsx
@@ -1,20 +1,21 @@
+"use client";
+
 import {
   CalendarDays,
   LayoutDashboard,
+  LogOut,
   Stethoscope,
   UsersRound,
 } from "lucide-react";
-import { headers } from "next/headers";
 import Image from "next/image";
 import Link from "next/link";
+import { usePathname, useRouter } from "next/navigation";
 
-import LogoutButton from "@/app/(auth)/components/logout-button";
 import { Avatar, AvatarFallback } from "@/components/ui/avatar";
 import {
   DropdownMenu,
   DropdownMenuContent,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
+  DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import {
@@ -29,7 +30,7 @@ import {
   SidebarMenuButton,
   SidebarMenuItem,
 } from "@/components/ui/sidebar";
-import { auth } from "@/lib/auth";
+import { authClient } from "@/lib/auth-client";
 
 const items = [
   {
@@ -54,20 +55,25 @@ const items = [
   },
 ];
 
-export async function AppSidebar() {
-  const session = await auth.api.getSession({
-    headers: await headers(),
-  });
+export function AppSidebar() {
+  const router = useRouter();
+  const session = authClient.useSession();
+  const pathname = usePathname();
 
-
+  const handleSignOut = async () => {
+    await authClient.signOut({
+      fetchOptions: {
+        onSuccess: () => {
+          router.push("/authentication");
+        },
+      },
+    });
+  };
   return (
     <Sidebar>
-      <SidebarHeader className="p-4">
-        <Link href="/dashboard">
-          <Image src="/logo.svg" alt="Dr agenda logo" width={136} height={28} />
-        </Link>
+      <SidebarHeader className="border-b p-4">
+        <Image src="/logo.svg" alt="Doutor Agenda" width={136} height={28} />
       </SidebarHeader>
-
       <SidebarContent>
         <SidebarGroup>
           <SidebarGroupLabel>Menu Principal</SidebarGroupLabel>
@@ -75,7 +81,7 @@ export async function AppSidebar() {
             <SidebarMenu>
               {items.map((item) => (
                 <SidebarMenuItem key={item.title}>
-                  <SidebarMenuButton asChild>
+                  <SidebarMenuButton asChild isActive={pathname === item.url}>
                     <Link href={item.url}>
                       <item.icon />
                       <span>{item.title}</span>
@@ -87,7 +93,6 @@ export async function AppSidebar() {
           </SidebarGroupContent>
         </SidebarGroup>
       </SidebarContent>
-
       <SidebarFooter>
         <SidebarMenu>
           <SidebarMenuItem>
@@ -97,22 +102,21 @@ export async function AppSidebar() {
                   <Avatar>
                     <AvatarFallback>F</AvatarFallback>
                   </Avatar>
-
-                  <div className="flex flex-col justify-center text-sm">
-                    <p className="font-semibold">
-                      {session?.user?.clinic?.name}
+                  <div>
+                    <p className="text-sm">
+                      {session.data?.user?.clinic?.name}
                     </p>
-                    <p className="text-muted-foreground ">
-                      {session?.user.email}
+                    <p className="text-muted-foreground text-sm">
+                      {session.data?.user.email}
                     </p>
                   </div>
                 </SidebarMenuButton>
               </DropdownMenuTrigger>
-
               <DropdownMenuContent>
-                <DropdownMenuLabel>My Account</DropdownMenuLabel>
-                <DropdownMenuSeparator />
-                <LogoutButton />
+                <DropdownMenuItem onClick={handleSignOut}>
+                  <LogOut />
+                  Sair
+                </DropdownMenuItem>
               </DropdownMenuContent>
             </DropdownMenu>
           </SidebarMenuItem>
